Code edits / Added.
Added: New boolean method to validate if the string is a number only without being empty or whitespaced.
Edited: SQLiteconnection conn method to just have sqlite_conn as there only is one database that is being worked upon.
Edited: Consolewritelines with "-" to become stringbuilder.
Added: A data connection (once) to fill up lists with information from the database if any. If it fails, we just have empty lists.